pyshiny/requirements_no-deps.txt
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
comm==0.1.3
contextvars==2.4
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
h11==0.14.0
html2image==2.0.2
htmltools==0.2.1
idna==3.4
img2pdf==0.4.4
immutables==0.19
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaleido==0.2.1
linkify-it-py==2.0.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathlib==1.0.1
pdfminer==20191125
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.0.0
Pillow==9.4.0
platformdirs==3.8.1
plotly==5.15.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.15.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pyzmq==25.1.0
requests==2.28.2
shiny==0.4.0
shinyswatch==0.2.4
shinywidgets==0.2.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.30.0
tenacity==8.2.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.2.3
websockets==11.0.3
widgetsnbextension==4.0.8
xlrd==2.0.1
XStatic-bootswatch==3.3.7.0
zipp==3.16.1
